Reorganized code into folders with individual CSS files. Renamed the updateStudentArrays function.
