+++ src/components/Camp/camp.css
@@ -1,0 +1,16 @@
+.camp-container {
+  width: 50vw;
+  display: flex;
+  flex-direction: column;
+}
+
+.camp {
+  width: 100%;
+  height: 100vh;
+  display: flex;
+  flex-flow: row wrap;
+  text-align: center;
+  margin: auto;
+  justify-content: center;
+  align-content: flex-start;
+}
