+++ src/components/Camps/Camps.js
@@ -1,14 +1,15 @@
 import React from 'react';
-import Camp from './Camp';
+import Camp from '../Camp/Camp';
+import './camps.css'
 
-const Camps = ({ hogwarts, voldemorts, setHouses, studentsCopy, setSingleStudent }) => {
+const Camps = ({ hogwarts, voldemorts, updateStudentArrays, studentsCopy, setSingleStudent }) => {
   return (<div className="camps">
     <Camp
       arr={hogwarts}
       className={'hogwarts'}
       heading={'Hogwarts Students'}
       key={'hogwarts'}
-      setHouses={setHouses}
+      updateStudentArrays={updateStudentArrays}
       studentsCopy={studentsCopy}
       setSingleStudent={setSingleStudent}
     />
@@ -17,7 +18,7 @@ const Camps = ({ hogwarts, voldemorts, setHouses, studentsCopy, setSingleStudent
       className={'voldemorts'}
       heading={`Voldemort's Army`}
       key={'voldemorts'}
-      setHouses={setHouses}
+      updateStudentArrays={updateStudentArrays}
       studentsCopy={studentsCopy}
       setSingleStudent={setSingleStudent}
     />
