+++ src/components/Camps/camps.css
@@ -1,0 +1,4 @@
+.camps {
+  display: flex;
+  justify-content: center;
+}
