+++ src/components/Card/CardImage.js
@@ -1,9 +1,9 @@
 import React from 'react'
-import Gryffindor from '../media/Gryffindor.png';
-import Slytherin from '../media/Slytherin.png';
-import Ravenclaw from '../media/Ravenclaw.png';
-import Hufflepuff from '../media/Hufflepuff.png';
-import deaathEater from '../media/deathEater.png'
+import Gryffindor from '../../media/Gryffindor.png';
+import Slytherin from '../../media/Slytherin.png';
+import Ravenclaw from '../../media/Ravenclaw.png';
+import Hufflepuff from '../../media/Hufflepuff.png';
+import deaathEater from '../../media/deathEater.png'
 
 const CardImage = ({ camp, house }) =>
   camp === 'voldemorts' ? <img src={deaathEater} alt='Death Eather crest' className='card-image' />
