+++ src/components/SortButtons/sort-buttons.css
@@ -1,0 +1,7 @@
+.sort-buttons {
+  border: solid 1px #181818;
+}
+
+.sort-button {
+  width: 150px;
+}
