+++ src/helpers.js
@@ -5,3 +5,6 @@ export function rando(arr) {
 export const houses = [
   'Gryffindor', 'Slytherin', 'Ravenclaw', 'Hufflepuff'
 ]
+
+export const hogwartsStudents = (arr) => arr.filter(item => item.isStudent).sort((a, b) => a.house.localeCompare(b.house));
+export const voldemortsArmy = (arr) => arr.filter(item => !item.isStudent).sort((a, b) => a.name.localeCompare(b.name));
