+++ src/index.css
@@ -6,46 +6,12 @@ body {
   color: saddlebrown;
 }
 
-.welcome-btn {
-  background-color: #181818;
-  color: antiquewhite;
-  border: none;
-  border-radius: 5px;
-  padding: 5px 10px;
-  margin-top: 20px;
-  width: 20%;
-}
-
-.welcome-btn:hover {
-  background-color: #2c2c2c;
-}
-
-.welcome-hat {
-  margin-top: 50px;
-}
-
-.welcome-text {
-  width: 20%;
-  margin: 40px 40%;
-  font-size: 30px;
-  text-shadow: 0 0 1px black;
-}
-
 .heading {
   font-family: 'Cinzel', serif;
   font-size: 2.5em;
   margin-top: 100px;
   margin-bottom: 40px;
   text-shadow: 0 0 1px black;
-}
-
-.new-student {
-  display: flex;
-  justify-content: center;
-}
-
-.add-student {
-  margin-bottom: 30px;
 }
 
 .student-name,
@@ -61,81 +27,6 @@ body {
   background-color: beige;
   border: 1px solid #181818;
   box-shadow: none;
-}
-
-.add-btn {
-  background-color: #181818;
-  color: antiquewhite;
-}
-
-.add-btn:hover {
-  background-color: #2c2c2c;
-  color: antiquewhite;
-}
-
-.camps {
-  display: flex;
-  justify-content: center;
-}
-
-.camp-container {
-  width: 50vw;
-  display: flex;
-  flex-direction: column;
-}
-
-.camp {
-  width: 100%;
-  height: 100vh;
-  display: flex;
-  flex-flow: row wrap;
-  text-align: center;
-  margin: auto;
-  justify-content: center;
-  align-content: flex-start;
-}
-
-.card {
-  border: none;
-  width: 30%;
-  margin: 12px;
-  height: 15em;
-}
-
-.card-body {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-}
-
-.card-img-container {
-  background-color: rosybrown;
-  border-top: solid 1px #181818;
-  border-left: solid 1px #181818;
-  border-right: solid 1px #181818;
-  border-top-right-radius: 10px;
-  border-top-left-radius: 10px;
-  padding: 15px;
-}
-
-.card-image {
-  height: 100px;
-}
-
-.card-body {
-  border-bottom: solid 1px #181818;
-  border-left: solid 1px #181818;
-  border-right: solid 1px #181818;
-  border-bottom-right-radius: 10px;
-  border-bottom-left-radius: 10px;
-}
-
-.sort-buttons {
-  border: solid 1px #181818;
-}
-
-.sort-button {
-  width: 150px;
 }
 
 .btn-dark,
@@ -158,33 +49,7 @@ body {
   color: beige;
 }
 
-/* GRYFFINDOR RED #740001 GOLD #D3A625
-SLYTHERIN GREEN #1A472A SILVER ##898989
-RAVENCLAW BLUE #0E1A40 BRONZE #946B2D
-HUFFLEPUFF YELLOW #FFD800 BLACK #000000 */
-
-.Gryffindor-card {
-  background-color: #D3A625;
-  color: #740001;
-}
-
-.Slytherin-card {
-  background-color: #898989;
-  color: #1A472A;
-}
-
-.Ravenclaw-card {
-  background-color: #946B2D;
-  color: #0E1A40;
-}
-
-.Hufflepuff-card {
-  background-color: #000000;
-  color: #FFD800;
-}
-
-.Gryffindor-btn,
-.Gryffindor-img-container {
+.Gryffindor-btn {
   background-color: #740001;
   color: #D3A625;
 }
@@ -196,8 +61,7 @@ HUFFLEPUFF YELLOW #FFD800 BLACK #000000 */
   color: #D3A625;
 }
 
-.Slytherin-btn,
-.Slytherin-img-container {
+.Slytherin-btn {
   background-color: #1A472A;
   color: #898989;
 }
@@ -209,8 +73,7 @@ HUFFLEPUFF YELLOW #FFD800 BLACK #000000 */
   color: #898989;
 }
 
-.Ravenclaw-btn,
-.Ravenclaw-img-container {
+.Ravenclaw-btn {
   background-color: #0E1A40;
   color: #946B2D;
 }
@@ -222,8 +85,7 @@ HUFFLEPUFF YELLOW #FFD800 BLACK #000000 */
   color: #946B2D;
 }
 
-.Hufflepuff-btn,
-.Hufflepuff-img-container {
+.Hufflepuff-btn {
   background-color: #FFD800;
   color: #000000;
 }
@@ -233,27 +95,4 @@ HUFFLEPUFF YELLOW #FFD800 BLACK #000000 */
 .Hufflepuff-btn:focus {
   background-color: #d0a22e;
   color: #000000;
-}
-
-.voldemort-card {
-  background-color: beige;
-}
-
-.voldemort-img-container {
-  background-color: #4d5c3c;
-}
-
-.modal-content {
-  background-image: url(./media/paper.png);
-}
-
-.modal-header {
-  color: #181818;
-  font-size: 1.25rem;
-  text-shadow: 0 0 1px black;
-}
-
-.modal-header,
-.modal-footer {
-  border: none;
 }
